+++ app/adapters/citation.ts
@@ -1,0 +1,11 @@
+import OsfAdapter from './osf-adapter';
+
+export default class Citation extends OsfAdapter {
+}
+
+
+declare module 'ember-data' {
+    interface AdapterRegistry {
+        'citation': Citation;
+    }
+}