+++ app/adapters/comment.ts
@@ -1,0 +1,11 @@
+import OsfAdapter from './osf-adapter';
+
+export default class Comment extends OsfAdapter {
+}
+
+
+declare module 'ember-data' {
+    interface AdapterRegistry {
+        'comment': Comment;
+    }
+}