+++ app/adapters/file-contents.ts
@@ -1,0 +1,11 @@
+import OsfAdapter from './osf-adapter';
+
+export default class FileContents extends OsfAdapter {
+}
+
+
+declare module 'ember-data' {
+    interface AdapterRegistry {
+        'file-contents': FileContents;
+    }
+}