+++ app/adapters/file-provider.ts
@@ -1,0 +1,14 @@
+import OsfAdapter from './osf-adapter';
+
+export default class FileProvider extends OsfAdapter {
+    pathForType(): string {
+        return 'files';
+    }
+}
+
+
+declare module 'ember-data' {
+    interface AdapterRegistry {
+      'file-provider': FileProvider;
+    }
+}