+++ app/adapters/taxonomy.ts
@@ -1,0 +1,11 @@
+import OsfAdapter from './osf-adapter';
+
+export default class Taxonomy extends OsfAdapter {
+}
+
+
+declare module 'ember-data' {
+    interface AdapterRegistry {
+        'taxonomy': Taxonomy;
+    }
+}