+++ app/serializers/comment-report.ts
@@ -1,0 +1,10 @@
+import OsfSerializer from './osf-serializer';
+
+export default class CommentReport extends OsfSerializer {}
+
+
+declare module 'ember-data' {
+    interface SerializerRegistry {
+        'comment-report': CommentReport;
+    }
+}