+++ app/serializers/file-provider.js
@@ -1,7 +1,0 @@
-import OsfSerializer from './osf-serializer';
-
-export default OsfSerializer.extend({
-    modelNameFromPayloadKey() {
-        return 'file-provider';
-    },
-});