+++ app/serializers/preprint.ts
@@ -1,0 +1,24 @@
+import DS from 'ember-data'; // eslint-disable-line no-unused-vars
+import OsfSerializer from './osf-serializer';
+
+export default class Preprint extends OsfSerializer.extend({
+    serialize(snapshot: DS.Snapshot): any {
+        const res: any = this._super(...arguments);
+        if (res.data.attributes && 'subjects' in snapshot.record.changedAttributes()) { res.data.attributes.subjects = (snapshot.record.get('subjects') || []); }
+        return res;
+    },
+}) {
+    relationshipTypes = {
+        primaryFile: 'files',
+        node: 'nodes',
+        provider: 'providers',
+        license: 'licenses',
+    };
+}
+
+
+declare module 'ember-data' {
+    interface SerializerRegistry {
+        'preprint': Preprint;
+    }
+}